requirements.txt
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.40.0
django-cors-headers==3.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
flex==6.14.1
idna==2.8
importlib-metadata==1.2.0
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
jsonpointer==2.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
oauthlib==3.1.0
packaging==19.2
pip-autoremove==0.9.1
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.2
pyrsistent==0.15.6
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.12.0
sqlparse==0.3.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
validate-email==1.3
wrapt==1.11.2
zipp==0.6.0
